# Remove dup from Request#set_cotent_type