--- a/actionpack/lib/action_dispatch/http/response.rb
+++ b/actionpack/lib/action_dispatch/http/response.rb
@@ -448,7 +448,7 @@ module ActionDispatch # :nodoc:
     end
 
     def set_content_type(content_type, charset)
-      type = (content_type || "").dup
+      type = +(content_type || "")
       type << "; charset=#{charset.to_s.downcase}" if charset
       set_header CONTENT_TYPE, type
     end