@matthewd 2020-05-26
Is it really guaranteed that `set_content_type` will never be passed a mutable string that came from somewhere else? :confused:

This looks like it will mutate strings that might belong to someone else.