@kamipo 2020-05-26
Since `set_content_type` is private API, end users should use `#content_type=` if set a content type, I think.