@matthewd 2020-05-26
Ah okay, are all the internal callers guaranteed to discard the argument without using it after they pass it in here? If so I guess that's fine, though seems fragile for future changes 👍🏻🤷🏻‍♂️