@kamipo 2020-05-26
I agree that it is fragile for future changes, so I've changed to string interpolation in 7c936996a40291b71bdeeee7a3de1aa32bc4a4d8.